JDK 8 time support
This implements support for JDK 8 date and time API based on Joda Time.

The Calendar format is now based on the `DateTimeFormatter` instead of JAXB's `DatatypeConverter` that is no longer a part of the core modules in JDK 9.

Because `ZonedDateTime` contains `ZoneId`, the output is no longer ISO 8601 compilant in that it appends weird `[UTC]`. This is because ISO 8601 captures an instant of time whereas `ZonedDateTime` is capturing an instant + local rules about summer time etc.