Include content hash into File hash value
Problem
-------
I'd like to reuse JsonFormat structure to calculate hash of an arbitrary
value. This idea was already implemented in #26, but I'd like File's
hash value to include the content hash.

Solution
--------
This changes the default JsonFormat of the File to include both the file
name and the content hash.